requirements.txt
Flask
Flask-SQLAlchemy
reportlab
psycopg2-binary
python-escpos
Pillow